requirements.txt
streamlit
Pillow
google-cloud-vision
google-auth
opencv-python-headless
numpy
google-cloud-aiplatform 
matplotlib
seaborn